# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/panel/pay-requests": {
            "get": {
                "tags": [
                    "Pay Requests"
                ],
                "summary": "load gateways List with applying filter",
                "description": "Pay requests list",
                "operationId": "7f54f8473f1c6e4fa440212e4a7ae351",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "order_code",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "pay_requests": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "code": {
                                                                "type": "integer",
                                                                "example": 102
                                                            },
                                                            "gateway": {
                                                                "type": "string",
                                                                "example": "Shepa"
                                                            },
                                                            "amount": {
                                                                "type": "integer",
                                                                "example": 1500000
                                                            },
                                                            "order_code": {
                                                                "type": "integer",
                                                                "example": 124211
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "status": {
                                                                "properties": {
                                                                    "key": {
                                                                        "type": "string",
                                                                        "example": "Pending"
                                                                    },
                                                                    "label": {
                                                                        "type": "string",
                                                                        "example": "درانتظار پرداخت"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "paginate": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "from": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "to": {
                                                            "type": "integer",
                                                            "example": 2
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}